refactored POllChart into its own component
decided to move PollChart into its own component because it requires a significant amount of boilerplate to style and it was starting to clutter PollDetail which only cares that the poll is displayed.
